refactor: rename FollowUpAnalysisTable parent directory
